refactor: rename .tsx files to match sidebar view names (#423)
Rename EditableJournalView to JournalView and HabitTracker to
HabitTrackerView for consistent naming with sidebar navigation.
Remove orphaned DayView component that was only referenced by its
own test file.
