--- frontend/src/components/bujo/DayView.tsx
@@ -1,286 +1,0 @@
-import { DayEntries, Entry } from '@/types/bujo';
-import { EntryItem } from './EntryItem';
-import { Calendar } from 'lucide-react';
-import { format, isToday, isTomorrow, isYesterday } from 'date-fns';
-import { cn } from '@/lib/utils';
-import { useState, useMemo, useEffect } from 'react';
-import { MarkEntryDone, MarkEntryUndone, CancelEntry, UncancelEntry, CyclePriority, RetypeEntry, MoveEntryToRoot } from '@/wailsjs/go/wails/App';
-
-interface DayViewProps {
-  day: DayEntries;
-  selectedEntryId?: number | null;
-  onEntryChanged?: () => void;
-  onSelectEntry?: (id: number) => void;
-  onEditEntry?: (entry: Entry) => void;
-  onDeleteEntry?: (entry: Entry) => void;
-  onMigrateEntry?: (entry: Entry) => void;
-  onAddChild?: (entry: Entry) => void;
-  onAnswerEntry?: (entry: Entry) => void;
-  onMoveToList?: (entry: Entry) => void;
-}
-
-function buildTree(entries: Entry[]): Entry[] {
-  const map = new Map<number, Entry>();
-  const roots: Entry[] = [];
-  
-  entries.forEach(e => {
-    map.set(e.id, { ...e, children: [] });
-  });
-  
-  entries.forEach(e => {
-    const entry = map.get(e.id)!;
-    if (e.parentId && map.has(e.parentId)) {
-      map.get(e.parentId)!.children!.push(entry);
-    } else {
-      roots.push(entry);
-    }
-  });
-  
-  return roots;
-}
-
-function formatDayLabel(dateStr: string): string {
-  const date = new Date(dateStr + 'T00:00:00');
-  if (isToday(date)) return 'Today';
-  if (isTomorrow(date)) return 'Tomorrow';
-  if (isYesterday(date)) return 'Yesterday';
-  return format(date, 'EEEE, MMM d');
-}
-
-interface EntryTreeProps {
-  entries: Entry[];
-  depth?: number;
-  collapsedIds: Set<number>;
-  selectedEntryId?: number | null;
-  onToggleCollapse: (id: number) => void;
-  onToggleDone: (id: number) => void;
-  onSelect?: (id: number) => void;
-  onEdit?: (entry: Entry) => void;
-  onDelete?: (entry: Entry) => void;
-  onCancel?: (entry: Entry) => void;
-  onUncancel?: (entry: Entry) => void;
-  onCyclePriority?: (entry: Entry) => void;
-  onMigrate?: (entry: Entry) => void;
-  onCycleType?: (entry: Entry) => void;
-  onAddChild?: (entry: Entry) => void;
-  onMoveToRoot?: (entry: Entry) => void;
-  onMoveToList?: (entry: Entry) => void;
-  onAnswer?: (entry: Entry) => void;
-  showContextDot?: boolean;
-}
-
-function EntryTree({ entries, depth = 0, collapsedIds, selectedEntryId, onToggleCollapse, onToggleDone, onSelect, onEdit, onDelete, onCancel, onUncancel, onCyclePriority, onMigrate, onCycleType, onAddChild, onMoveToRoot, onMoveToList, onAnswer, showContextDot = true }: EntryTreeProps) {
-  return (
-    <>
-      {entries.map((entry) => {
-        const hasChildren = entry.children && entry.children.length > 0;
-        const isCollapsed = collapsedIds.has(entry.id);
-
-        return (
-          <div key={entry.id}>
-            <EntryItem
-              entry={entry}
-              depth={depth}
-              isCollapsed={isCollapsed}
-              hasChildren={hasChildren}
-              hasParent={entry.parentId !== null}
-              childCount={entry.children?.length || 0}
-              isSelected={entry.id === selectedEntryId}
-              onToggleCollapse={() => onToggleCollapse(entry.id)}
-              onToggleDone={() => onToggleDone(entry.id)}
-              onSelect={onSelect ? () => onSelect(entry.id) : undefined}
-              onEdit={onEdit ? () => onEdit(entry) : undefined}
-              onDelete={onDelete ? () => onDelete(entry) : undefined}
-              onCancel={onCancel ? () => onCancel(entry) : undefined}
-              onUncancel={onUncancel ? () => onUncancel(entry) : undefined}
-              onCyclePriority={onCyclePriority ? () => onCyclePriority(entry) : undefined}
-              onMigrate={onMigrate ? () => onMigrate(entry) : undefined}
-              onCycleType={onCycleType ? () => onCycleType(entry) : undefined}
-              onAddChild={onAddChild && entry.type !== 'question' ? () => onAddChild(entry) : undefined}
-              onMoveToRoot={onMoveToRoot ? () => onMoveToRoot(entry) : undefined}
-              onMoveToList={onMoveToList ? () => onMoveToList(entry) : undefined}
-              onAnswer={onAnswer && entry.type === 'question' ? () => onAnswer(entry) : undefined}
-              showContextDot={showContextDot}
-            />
-            {hasChildren && !isCollapsed && (
-              <EntryTree
-                entries={entry.children!}
-                depth={depth + 1}
-                collapsedIds={collapsedIds}
-                selectedEntryId={selectedEntryId}
-                onToggleCollapse={onToggleCollapse}
-                onToggleDone={onToggleDone}
-                onSelect={onSelect}
-                onEdit={onEdit}
-                onDelete={onDelete}
-                onCancel={onCancel}
-                onUncancel={onUncancel}
-                onCyclePriority={onCyclePriority}
-                onMigrate={onMigrate}
-                onCycleType={onCycleType}
-                onAddChild={onAddChild}
-                onMoveToRoot={onMoveToRoot}
-                onMoveToList={onMoveToList}
-                onAnswer={onAnswer}
-                showContextDot={showContextDot}
-              />
-            )}
-          </div>
-        );
-      })}
-    </>
-  );
-}
-
-function getParentIds(tree: Entry[]): Set<number> {
-  const parentIds = new Set<number>();
-  const visit = (entries: Entry[]) => {
-    entries.forEach(entry => {
-      if (entry.children && entry.children.length > 0) {
-        parentIds.add(entry.id);
-        visit(entry.children);
-      }
-    });
-  };
-  visit(tree);
-  return parentIds;
-}
-
-export function DayView({ day, selectedEntryId, onEntryChanged, onSelectEntry, onEditEntry, onDeleteEntry, onMigrateEntry, onAddChild, onAnswerEntry, onMoveToList }: DayViewProps) {
-  const tree = useMemo(() => buildTree(day.entries), [day.entries]);
-  const initialCollapsed = useMemo(() => getParentIds(tree), [tree]);
-  const [collapsedIds, setCollapsedIds] = useState<Set<number>>(initialCollapsed);
-  const dateObj = new Date(day.date + 'T00:00:00');
-
-  useEffect(() => {
-    setCollapsedIds(getParentIds(tree));
-  }, [tree]);
-
-  const toggleCollapse = (id: number) => {
-    setCollapsedIds(prev => {
-      const next = new Set(prev);
-      if (next.has(id)) {
-        next.delete(id);
-      } else {
-        next.add(id);
-      }
-      return next;
-    });
-  };
-
-  const handleToggleDone = async (id: number) => {
-    const entry = day.entries.find(e => e.id === id);
-    if (!entry || (entry.type !== 'task' && entry.type !== 'done')) return;
-
-    try {
-      if (entry.type === 'done') {
-        await MarkEntryUndone(id);
-      } else {
-        await MarkEntryDone(id);
-      }
-      onEntryChanged?.();
-    } catch (error) {
-      console.error('Failed to toggle entry:', error);
-    }
-  };
-
-  const handleCancelEntry = async (entry: Entry) => {
-    try {
-      await CancelEntry(entry.id);
-      onEntryChanged?.();
-    } catch (error) {
-      console.error('Failed to cancel entry:', error);
-    }
-  };
-
-  const handleUncancelEntry = async (entry: Entry) => {
-    try {
-      await UncancelEntry(entry.id);
-      onEntryChanged?.();
-    } catch (error) {
-      console.error('Failed to uncancel entry:', error);
-    }
-  };
-
-  const handleCyclePriority = async (entry: Entry) => {
-    try {
-      await CyclePriority(entry.id);
-      onEntryChanged?.();
-    } catch (error) {
-      console.error('Failed to cycle priority:', error);
-    }
-  };
-
-  const handleCycleType = async (entry: Entry) => {
-    const cycleOrder = ['task', 'note', 'event', 'question'] as const;
-    const currentIndex = cycleOrder.indexOf(entry.type as typeof cycleOrder[number]);
-    if (currentIndex === -1) return;
-    const nextType = cycleOrder[(currentIndex + 1) % cycleOrder.length];
-    try {
-      await RetypeEntry(entry.id, nextType);
-      onEntryChanged?.();
-    } catch (error) {
-      console.error('Failed to cycle type:', error);
-    }
-  };
-
-  const handleMoveToRoot = async (entry: Entry) => {
-    try {
-      await MoveEntryToRoot(entry.id);
-      onEntryChanged?.();
-    } catch (error) {
-      console.error('Failed to move entry to root:', error);
-    }
-  };
-
-
-  return (
-    <div className="animate-fade-in">
-      {/* Day Header */}
-      <div className="flex items-center gap-4 mb-3 pb-2 border-b border-border">
-        <div className="flex items-center gap-2">
-          <Calendar className="w-4 h-4 text-primary" />
-          <h3 className={cn(
-            'font-display text-lg font-semibold',
-            isToday(dateObj) && 'text-primary'
-          )}>
-            {formatDayLabel(day.date)}
-          </h3>
-          <span className="text-xs text-muted-foreground">
-            {format(dateObj, 'MMM d, yyyy')}
-          </span>
-        </div>
-      </div>
-
-      {/* Entries */}
-      <div className="space-y-0.5">
-        {tree.length > 0 ? (
-          <EntryTree
-            entries={tree}
-            collapsedIds={collapsedIds}
-            selectedEntryId={selectedEntryId}
-            onToggleCollapse={toggleCollapse}
-            onToggleDone={handleToggleDone}
-            onSelect={onSelectEntry}
-            onEdit={onEditEntry}
-            onDelete={onDeleteEntry}
-            onCancel={handleCancelEntry}
-            onUncancel={handleUncancelEntry}
-            onCyclePriority={handleCyclePriority}
-            onMigrate={onMigrateEntry}
-            onCycleType={handleCycleType}
-            onAddChild={onAddChild}
-            onMoveToRoot={handleMoveToRoot}
-            onMoveToList={onMoveToList}
-            onAnswer={onAnswerEntry}
-            showContextDot={false}
-          />
-        ) : (
-          <p className="text-sm text-muted-foreground italic py-4 text-center">
-            No entries yet. Start journaling!
-          </p>
-        )}
-      </div>
-    </div>
-  );
-}
