+++ frontend/src/components/bujo/HabitTrackerView.core.test.tsx
@@ -1,7 +1,7 @@
 import { describe, it, expect, vi, beforeEach } from 'vitest'
 import { render, screen, waitFor, fireEvent } from '@testing-library/react'
 import userEvent from '@testing-library/user-event'
-import { HabitTracker } from './HabitTracker'
+import { HabitTracker } from './HabitTrackerView'
 import { Habit } from '@/types/bujo'
 
 vi.mock('@/wailsjs/go/wails/App', () => ({
