+++ frontend/src/components/bujo/HabitTrackerView.goals.test.tsx
@@ -1,7 +1,7 @@
 import { describe, it, expect, vi, beforeEach, afterEach } from 'vitest'
 import { render, screen, waitFor } from '@testing-library/react'
 import userEvent from '@testing-library/user-event'
-import { HabitTracker } from './HabitTracker'
+import { HabitTracker } from './HabitTrackerView'
 import { Habit } from '@/types/bujo'
 
 vi.mock('@/wailsjs/go/wails/App', () => ({
